% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cens}
\alias{cens}
\title{Event indicator vector cens}
\format{
\subsection{\code{cens}}{

A vector of length 100 of event indicators.
}
}
\usage{
cens
}
\description{
This vector contains binary indicators reflecting whether or not a sample
had the event. If the vector equals 1, this indicates the corresponding
patient had the event and is 0 otherwise.
}
\keyword{datasets}
