% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1.df}
\alias{data1.df}
\title{Simulated data 1}
\format{
\subsection{\code{data1.df}}{

A data frame with 29951 rows and 4 columns:
\describe{
\item{PID}{Patient ID}
\item{id}{Image ID (which matches Patient ID)}
\item{x}{x-coordinate of each cell}
\item{y}{y-coordinate of each cell}
}
}
}
\usage{
data1.df
}
\description{
This dataset contains simulated data. The images were split into two groups.
The first group was simulated to have large squares. The second group was
simulated to have loops. This dataset contains 100 samples.
}
\keyword{datasets}
