% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data2.df}
\alias{data2.df}
\title{Simulated data 2}
\format{
\subsection{\code{data2.df}}{

A data frame with 100481 rows and 4 columns:
\describe{
\item{PID}{Patient ID}
\item{id}{Image ID (which matches Patient ID)}
\item{x}{x-coordinate of each cell}
\item{y}{y-coordinate of each cell}
}
}
}
\usage{
data2.df
}
\description{
This dataset contains simulated data. The images were simulated using the
\code{scSpatialSIM} package to represent different tissue structures. Specifically,
we used \code{scSpatialSIM} to simulate two different tissues and separated these
into different images. This dataset contains 100 samples.
}
\keyword{datasets}
