% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_importance.R
\name{generate_rips}
\alias{generate_rips}
\title{Generate a list of persistence diagrams}
\usage{
generate_rips(images.df, max.threshold, print.progress = TRUE)
}
\arguments{
\item{images.df}{A data.frame containing the image information. See details.}

\item{max.threshold}{The maximum radius of the circles in the Rips filtration}

\item{print.progress}{Boolean indicating whether progress in constructing the
Rips filtrations across images should be printed.}
}
\value{
Returns a list of persistence diagrams for each image.
}
\description{
Construct a Rips filtration for each image and return a list of persistence diagrams.
This is a helper function for \code{scale_importance}.
}
\examples{
# Generate a persistence diagram based on a Rips filtration for each image
pd.list <- generate_rips(data1.df, 100)
}
