% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_connections}
\alias{plot_cell_connections}
\title{Heatmap of cell-cell connections}
\usage{
plot_cell_connections(image, threshold, title, type.column, unique.types)
}
\arguments{
\item{image}{A data.frame containing the image information.}

\item{threshold}{Radius to plot, usually \code{threshold} generated by \code{scale_importance}}

\item{title}{Plot title}

\item{type.column}{The name of the column in \code{images.df} that contains the cell types}

\item{unique.types}{Vector of unique cell types in the \code{type.column}}
}
\value{
Returns a ggplot of the heatmap of cell-cell connections.
}
\description{
Visualize the connectivity matrix produced by \code{generate_connectivity}
}
\examples{
require(magrittr)
plot_cell_connections(
   image = data1.df \%>\% dplyr::filter(PID == 1),
   threshold = 10,
   title = "",
   type.column = "type",
   unique.types = unique(data1.df$type)
)
}
