% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_persistence}
\alias{plot_persistence}
\title{Plot persistence diagrams}
\usage{
plot_persistence(PD, title, dims = c(100, 100))
}
\arguments{
\item{PD}{Persistence diagram for a given image}

\item{title}{Title for the output figure}

\item{dims}{Dimensions for the output figure}
}
\value{
Plot illustrating the birth and death scales for each homology
}
\description{
Plot persistence diagrams
}
\examples{
require(magrittr)
PD <- TDAstats::calculate_homology(
       data2.df \%>\% dplyr::filter(PID == 1) \%>\% dplyr::select(x,y),
       dim = 1,
       threshold = 10)
plot_persistence(PD = PD, title = "Example Persistence Diagram", dims = c(10, 10))
}
