% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopKAT.R
\name{rips_similarity_matrix}
\alias{rips_similarity_matrix}
\title{Calculate a similarity matrix based on the Rips filtration}
\usage{
rips_similarity_matrix(images.df, max.threshold, print.progress = TRUE)
}
\arguments{
\item{images.df}{A data.frame containing the image information. See details.}

\item{max.threshold}{The maximum radius of the circles in the Rips filtration}

\item{print.progress}{Boolean indicating whether progress in constructing the
Rips filtrations across images should be printed.}
}
\value{
Returns a list of similarity (kernel) matrices and a list of the
resulting persistence diagrams for each image.
}
\description{
Given input data, computes a Rips filtration and persistence diagram for
each image. Then constructs a pairwise similarity matrix for each homology group
(connected components and loops) based on the resulting persistence diagrams.
}
\details{
\code{images.df} should contain a \code{PID} column indicating which sample
each image corresponds to. It should also contain columns \code{x} and \code{y} indicating
the location of each cell. Each row in \code{images.df} corresponds to a cell within
each image. See the package vignettes for an example of the structure.

To choose \code{max.threshold}, a reasonable first choice is to select the maximum
possible radius that would cover all cells in each image. This could be the
maximum width of the largest image, for example.

This function should be run as a first step of the TopKAT analysis. After that,
the resulting kernel matrices may be fed into the \code{TopKAT} function.
}
\examples{
# Construct a Rips filtration for each simulated image
simmat <- rips_similarity_matrix(data1.df, max.threshold = 100, print.progress = TRUE)
}
