% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_importance.R
\name{scale_importance}
\alias{scale_importance}
\title{Quantifying the importance of various scale parameters}
\usage{
scale_importance(
  pd.list,
  y,
  X = NULL,
  cens = NULL,
  omega.list,
  threshold,
  PIDs,
  outcome.type = "continuous",
  method = "davies",
  omnibus = "cauchy",
  perm = FALSE,
  nperm = 999,
  n.thresh = 50,
  print.progress = FALSE
)
}
\arguments{
\item{pd.list}{List of persistence diagrams}

\item{y}{Outcome vector}

\item{X}{Covariates to adjust for, if desired. May be left NULL.}

\item{cens}{Censoring vector for survival outcomes. May be left NULL.}

\item{omega.list}{Vector of weights to combine kernel matrices}

\item{threshold}{Maximum radius for Rips filtration}

\item{PIDs}{Vector of patient IDs}

\item{outcome.type}{Outcome type, options include "continuous", "binary", or
"survival"}

\item{method}{Specifies how the kernel p-value should be calculated. Options
include "davies" for an analytical p-value calculation or "permutation". Only
for continuous or binary outcomes.}

\item{omnibus}{If providing multiple kernel matrices, specify how the p-values
should be combined. Defaults to "cauchy" for the Cauchy combination test.}

\item{perm}{Required only for survival outcomes. Specifies if permutations should
be used. Default is FALSE and Davies method is used.}

\item{nperm}{If using permutations, how many permutations should be used?}

\item{n.thresh}{How many maximum radii should be considered?}

\item{print.progress}{Boolean, should progress be printed throughout analysis?}
}
\value{
Returns a list with the following elements:
\describe{
\item{min.thresh}{The radius value at which the lowest TopKAT p-value was obtained}
\item{pvals}{The vector of TopKAT p-values for each radius value}
\item{threshold.seq}{The vector of maximum radii considered}
}
}
\description{
Identify which radii are most important in characterizing an association
between the topological structure in each image and outcomes.
}
\details{
The arguments for \code{method}, \code{omnibus}, \code{perm}, and \code{nperm} should match
what was used in the TopKAT analysis.
}
\examples{
# Generate a persistence diagram based on a Rips filtration for each image
pd.list <- generate_rips(data1.df, 100)
# Run the scale importance analysis
data1.scale <- scale_importance(pd.list = pd.list,
  y = y,
  cens = cens,
  omega.list = c(0, 0.5, 1),
  threshold = 100,
  PIDs = 1:100,
  outcome.type = "survival")

# Plot the results
plot(data1.scale$threshold.seq, data1.scale$pvals); abline(v = data1.scale$min.thresh)
}
