% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopKAT.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Construct a pairwise similarity matrix}
\usage{
similarity_matrix(rips.list, n.sample, PIDs, print.progress = FALSE)
}
\arguments{
\item{rips.list}{A list of persistence diagrams for each sample}

\item{n.sample}{The number of samples in the dataset (the length of \code{rips.list})}

\item{PIDs}{Unique identifiers for each sample. This will be used to name the
resulting similarity matrix.}

\item{print.progress}{Should a progress bar be printed?}
}
\value{
Returns a list of kernel (similarity matrices) for each homology group.
}
\description{
\code{similarity_matrix} is a helper function used within the \code{rips_similarity_matrix}
function to construct a pairwise similarity matrix comparing pairs of
persistence diagrams.
}
\examples{
require(magrittr)
require(TDAstats)

# Save the PIDs
PIDs <- unique(data1.df$PID)

# Save the number of samples
n.sample <- length(PIDs)

# Initialize a list to store the Rips filtrations
rips.list <- lapply(1:n.sample, function(i) list()); names(rips.list) <- PIDs

# Iterate through the samples to construct Rips filtration
for (i in PIDs) {


 # Subset the data to just this PID
 data.i <- data1.df \%>\%
   dplyr::filter(PID == i) \%>\%
   dplyr::select(x,y)

 # Construct a Rips filtration using TDAstats
 rips.i <- TDAstats::calculate_homology(data.i, dim = 1, threshold = 10)

 # Save
 rips.list[[i]] <- rips.i
}
 K.list <- similarity_matrix(rips.list, n.sample, PIDs)
}
