% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopKAT.R
\name{TopKAT}
\alias{TopKAT}
\title{Topological Kernel Association Test (TopKAT)}
\usage{
TopKAT(
  y,
  X = NULL,
  cens = NULL,
  K.list,
  omega.list,
  outcome.type = "continuous"
)
}
\arguments{
\item{y}{Vector of outcomes. Must be numeric. For continuous and survival data,
should be a numeric vector. For a binary outcome, must consist of 0s and 1s.}

\item{X}{Matrix of covariates to adjust for. May be left as NULL.}

\item{cens}{Vector of event indicators for a survival outcome. 1 indicates a
sample experienced the event, 0 otherwise. If not using a survival outcome,
leave as NULL.}

\item{K.list}{List of 2 kernel matrices corresponding to similarities
among connected components and among loops. May provide only 1 kernel matrix if
interested in a specific homology.}

\item{omega.list}{Vector of weights to create different combinations of the kernel matrices.
Some suggested options: \code{c(0, 1)} then TopKAT will combine the p-values for a test associating the
connected components with y and the loops with y separately; \code{c(0, 0.5, 1)} will combine p-values
across just the connected components, an even split of connected components and loops,
and just loops.}

\item{outcome.type}{What kind of outcome is y? Options include "continuous",
"binary", or "survival"}
}
\value{
Returns a list of the following objects:
overall.pval: the overall p-value describing the association between similarities
in topological structures and clinical outcomes,
p.vals: the vector of individual p-values for each weight in omega.list,
y: the outcome provided,
X: the covariates provided,
omega.list: the vector of weights provided,
outcome.type: the outcome type specified
}
\description{
Perform global test of association between the geometric (topological)
structures of spatially-resolved images of cells with continuous, binary,
or survival outcomes.
}
\examples{
# First, construct the similarity matrix
simmat <- rips_similarity_matrix(data1.df, max.threshold = 100, print.progress = TRUE)

# Then, run TopKAT
res <- TopKAT(y = y,
              cens = cens,
              K.list = simmat$K.list,
              omega.list = c(0, 0.5, 1),
              outcome.type = "survival")

# Check result
res$overall.pval
}
