% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{Outcome vector y}
\format{
\subsection{\code{y}}{

A vector of length 100 of event times.
}
}
\usage{
y
}
\description{
This vector contains outcomes for the simulated data. These are simulated
survival outcomes from the exponential distribution. For 100 samples, we
simulated event times for two different exponential distributions.
}
\keyword{datasets}
